# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PortOS Public API",
    "version": "2.51.0",
    "description": "Complete registry contract for externally-callable PortOS services. Runtime exposure is opt-in per API in Settings → API Access."
  },
  "servers": [],
...
  "paths": {
    "/api/voice/public/synthesize": {
      "post": {
        "tags": [
          "voice"
        ],
        "summary": "Synthesize speech",
        "description": "Convert text to spoken audio (WAV) using the selected engine and voice.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 4000
                  },
                  "engine": {
                    "type": "string",
                    "enum": [
                      "kokoro",
                      "piper"
                    ]
                  },
                  "voice": {
                    "type": "string",
                    "maxLength": 128
                  },
                  "rate": {
                    "description": "Speech rate. Validated 0.25–4 (Piper). Kokoro clamps to 0.5–2.0.",
                    "type": "number",
                    "minimum": 0.25,
                    "maximum": 4
                  }
                },
                "required": [
                  "text"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "WAV audio",
            "content": {
              "audio/wav": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload or unknown voice"
          }
        },
        "x-portos-tool": {
          "name": "voice.synthesize",
          "version": 1,
          "policy": {
            "privacy": "personal",
            "sideEffect": "local-compute",
            "async": false
          }
        },
        "security": []
      }
    },
    "/api/voice/public/voices": {
      "get": {
        "tags": [
          "voice"
        ],
        "summary": "List voices",
        "description": "Enumerate available voices for an engine (query param `engine`, default = active).",
        "parameters": [
          {
            "name": "engine",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "kokoro",
                "piper"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Voice list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "engine",
                    "voices"
                  ],
                  "properties": {
                    "engine": {
                      "type": "string",
                      "enum": [
                        "kokoro",
                        "piper"
                      ]
                    },
                    "voices": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-portos-tool": {
          "name": "voice.list-voices",
          "version": 1,
          "policy": {
            "privacy": "internal",
            "sideEffect": "read",
            "async": false
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "PortOS session token (when the API requires auth)."
      },
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "PortOS password via HTTP Basic (username ignored)."
      }
    },
...
  }
}